+++ packages/angular/standalone/src/directives/searchbar.ts
@@ -7,26 +7,14 @@ import {
   HostListener,
   Injector,
   NgZone,
+  forwardRef,
 } from '@angular/core';
-import type { OnInit } from '@angular/core';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 import { ValueAccessor } from '@ionic/angular/common';
 import type { SearchbarInputEventDetail, SearchbarChangeEventDetail, Components } from '@ionic/core/components';
 import { defineCustomElement } from '@ionic/core/components/ion-searchbar.js';
 
-/**
- * Value accessor components should not use ProxyCmp
- * and should call defineCustomElement and proxyInputs
- * manually instead. Using both the @ProxyCmp and @Component
- * decorators and useExisting (where useExisting refers to the
- * class) causes ng-packagr to output multiple component variables
- * which breaks treeshaking.
- * For example, the following would be generated:
- * let IonSearchbar = IonSearchbar_1 = class IonSearchbar extends ValueAccessor {
- * Instead, we want only want the class generated:
- * class IonSearchbar extends ValueAccessor {
- */
-import { proxyInputs, proxyMethods, proxyOutputs } from './angular-component-lib/utils';
+import { ProxyCmp, proxyOutputs } from './angular-component-lib/utils';
 
 const SEARCHBAR_INPUTS = [
   'animated',
@@ -51,41 +39,41 @@ const SEARCHBAR_INPUTS = [
   'value',
 ];
 
-const SEARCHBAR_METHODS = ['setFocus', 'getInputElement'];
+/**
+ * Pulling the provider into an object and using PURE works
+ * around an ng-packagr issue that causes
+ * components with multiple decorators and
+ * a provider to be re-assigned. This re-assignment
+ * is not supported by Webpack and causes treeshaking
+ * to not work on these kinds of components.
+ */
+const accessorProvider = {
+  provide: NG_VALUE_ACCESSOR,
+  useExisting: /*@__PURE__*/ forwardRef(() => IonSearchbar),
+  multi: true,
+};
 
+@ProxyCmp({
+  defineCustomElementFn: defineCustomElement,
+  inputs: SEARCHBAR_INPUTS,
+  methods: ['setFocus', 'getInputElement'],
+})
 @Component({
   selector: 'ion-searchbar',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
   // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
   inputs: SEARCHBAR_INPUTS,
-  providers: [
-    {
-      provide: NG_VALUE_ACCESSOR,
-      useExisting: IonSearchbar,
-      multi: true,
-    },
-  ],
+  providers: [accessorProvider],
   standalone: true,
 })
-export class IonSearchbar extends ValueAccessor implements OnInit {
+export class IonSearchbar extends ValueAccessor {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone, injector: Injector) {
     super(injector, r);
-    defineCustomElement();
     c.detach();
     this.el = r.nativeElement;
     proxyOutputs(this, this.el, ['ionInput', 'ionChange', 'ionCancel', 'ionClear', 'ionBlur', 'ionFocus']);
-  }
-
-  ngOnInit(): void {
-    /**
-     * Data-bound input properties are set
-     * by Angular after the constructor, so
-     * we need to run the proxy in ngOnInit.
-     */
-    proxyInputs(IonSearchbar, SEARCHBAR_INPUTS);
-    proxyMethods(IonSearchbar, SEARCHBAR_METHODS);
   }
 
   @HostListener('ionInput', ['$event.target'])